authentication-service/requirements.txt
alembic==1.13.1
async-timeout==4.0.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Deprecated==1.2.14
Flask==2.0.1
Flask-And-Redis==1.0.0
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.0.1
jwcrypto==1.5.0
Mako==1.3.0
MarkupSafe==2.0.1
marshmallow==3.12.2
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2021.1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.21
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.0.1
wrapt==1.16.0
